refactor: fix rubocop warnings
