+++ lib/vkdonate/donate.rb
@@ -1,49 +1,34 @@
+# frozen_string_literal: true
+
 module Vkdonate
-  ##
   # Single donation
   class Donate
-    ##
-    # @return [Integer] Unique payment ID.
+    # @return [Integer] Unique payment ID
     attr_reader :id
 
-    ##
-    # @return [Integer] Donator VK ID.
+    # @return [Integer] Donator VK ID
     attr_reader :uid
 
-    ##
-    # @return [DateTime] Date and time of donation.
+    # @return [DateTime] Date and time of donation
     attr_reader :date
 
-    ##
-    # @return [Integer] Size of donation.
+    # @return [Integer] Size of donation
     attr_reader :sum
 
-    ##
-    # @return [String, nil] User message or +nil+ if empty.
+    # @return [String, nil] User message or +nil+ if empty
     attr_reader :msg
     alias message msg
 
-    ##
-    # @return [Boolean] Whether donation is anonymous.
+    # @return [Boolean] Whether donation is anonymous
     attr_reader :anon
     alias anon? anon
 
-    ##
-    # @return [Boolean] Whether donate is visible in group.
+    # @return [Boolean] Whether donate is visible in group
     attr_reader :visible
     alias visible? visible
 
-    ##
-    # @return [String] Human readable information.
-    def to_s
-      "Donation ##{id} by @#{uid} for #{sum}RUR (at #{date})"
-    end
-
-    ##
     # Save data about donation.
-    #
     # @param options [Hash]
-    #
     # @option options [Integer] id
     # @option options [Integer] uid
     # @option options [DateTime] date
@@ -69,11 +54,8 @@ module Vkdonate
       @visible = !!options[:visible]
     end
 
-    ##
     # Parse from JSON hash.
-    #
     # @param hash [Hash] parsed JSON hash.
-    #
     # @return [Donate]
     def self.from_json(hash)
       new(
@@ -85,6 +67,11 @@ module Vkdonate
         anon: !hash['anon'].to_i.zero?,
         visible: !hash['visible'].to_i.zero?
       )
+    end
+
+    # @return [String] Human readable information
+    def to_s
+      "Donation ##{id} by @#{uid} for #{sum}RUR (at #{date})"
     end
   end
 end
